+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.4b3 (unreleased)
 ------------------
 
+- Fix #64: Fails to startup w/o plone.app.iterate installed.
+  [jensens]
+
 - Reviewed and updated German translations.
   [jensens]